added like for details screen

+++ app/src/main/java/com/melyseev/cocktails2023/data/cache/dao/SubcategoriesDao.kt
@@ -12,8 +12,8 @@ interface SubcategoriesDao {
     suspend fun insert(entity: SubcategoryEntity)
 
     @Query("SELECT * from ${SubcategoryEntity.TABLE_SUBCATEGORY} WHERE ${SubcategoryEntity.ID_CATEGORY} = :idCategory")
-    suspend fun getAllSubcategories(idCategory: Long): List<SubcategoryEntity>
+    suspend fun getAllSubcategories(idCategory: Int): List<SubcategoryEntity>
 
     @Query("UPDATE ${SubcategoryEntity.TABLE_SUBCATEGORY} SET ${SubcategoryEntity.SUBCATEGORY_CHECKED} = :checkValue WHERE ${SubcategoryEntity.ID_CATEGORY} = :idCategory AND  ${SubcategoryEntity.SUBCATEGORY_NAME} = :subcategory")
-    suspend fun updateSubcategory(idCategory: Long, subcategory: String, checkValue: Int)
+    suspend fun updateSubcategory(idCategory: Int, subcategory: String, checkValue: Int)
 }
